requirements.txt
Flask==3.0.3
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.7
langchain_groq==0.2.0
langgraph==0.2.31
matplotlib==3.8.4
milvus_model==0.2.7
networkx==3.2.1
nltk==3.8.1
pandas==2.2.3
pydantic==2.9.2
pymilvus==2.4.7
pymssql==2.3.1
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.4.2
spacy==3.7.5
twilio==9.3.2
